\name{coauth}
\alias{coauth}
\title{coauth}
\usage{
  coauth(myname)
}
\arguments{
  \item{myname}{an author (length 2 character string)}
}
\value{
  the coauthors get the list of coauthors of the given name
}
\description{
  Return the coauthors of an author
}
\details{
  a helper function, mostly for Erdos example
}

