\name{getGroupID}
\alias{getGroupID}
\title{Get a Group ID from group name}
\usage{
  getGroupID(mc = NULL, group = NULL, ...,
    curl = getCurlHandle())
}
\arguments{
  \item{mc}{<what param does>}

  \item{group}{Name of group}

  \item{...}{<what param does>}

  \item{curl}{= getCurlHandle() <what param does>}
}
\value{
  numeric
}
\description{
  Meant for internal use.
}
\examples{
\dontrun{
getGroupID(mc, '')
}
}

