\name{MendeleyCredentials-class}
\alias{MendeleyCredentials-class}
\title{An S4 class that stores Mendeley credentials}
\description{
  An S4 class that stores Mendeley credentials

  An S4 class that stores Mendeley credentials

  An S4 class that stores MendeleyID

  An S4 class that stores MendeleyDocumentID
}

