\name{name_match}
\alias{name_match}
\title{name_match}
\usage{
  name_match(authors, target)
}
\arguments{
  \item{authors}{a list of authors, (length two character
  strings)}

  \item{target}{author (length 2 character string)}
}
\value{
  a list of the locations of a match
}
\description{
  match the full name of an author to the list
}
\details{
  a helper function, mostly for Erdos example
}

