\docType{package}
\name{RMendeley}
\alias{package-RMendeley}
\alias{RMendeley}
\alias{RMendeley-package}
\title{RMendeley - A programmatic interface to the Web Service methods provided by the Mendeley open science}
\description{
  This is version 1.0.0 of package RMendeley.
}
\details{
  \strong{Personal functions} \enumerate{ \item
  \link{mendeley_library} \item \link{mendeley_profile}
  \item \link{pubStats} \item \link{tagStats} \item
  \link{authoredPubs} \item \link{authorStats} \item
  \link{contacts} \item \link{addContact} }

  \strong{Folder functions} \enumerate{ \item
  \link{createFolder} \item \link{deleteFolder} \item
  \link{folderDocs} \item \link{Folders} \item
  \link{addFolderDoc} \item \link{getFolderID} }

  \strong{Document functions} \enumerate{ \item
  \link{docInfo} }

  \strong{Group functions} \enumerate{ \item
  \link{myGroups} \item \link{createGroup} \item
  \link{deleteGroup} \item \link{groupDocs} \item
  \link{groupDocInfo} \item \link{groupPeople} \item
  \link{leaveGroup} \item \link{unfollowGroup} \item
  \link{getGroupID} }

  \strong{Useful links} \enumerate{ \item
  \href{http://mendeley.com/}{Mendeley} \item
  \href{http://apidocs.mendeley.com/}{Full API
  documentation}}
}

