\docType{package}
\name{RMendeley}
\alias{package-RMendeley}
\alias{RMendeley}
\alias{RMendeley-package}
\title{RMendeley - Programmatic interface to the Mendeley Web}
\description{
  This is version 1.0.0 of package RMendeley.
  \strong{Personal functions} \link{mendeley_library}
  \link{mendeley_profile} \link{pubStats} \link{tagStats}
  \link{authoredPubs} \link{authorStats} \link{contacts}
  \link{addContact}
}
\details{
  \strong{Folder functions} \link{createFolder}
  \link{deleteFolder} \link{folderDocs} \link{Folders}
  \link{addFolderDoc} \link{getFolderID}

  \strong{Document functions} \link{docInfo}

  \strong{Group functions} \link{myGroups}
  \link{createGroup} \link{deleteGroup} \link{groupDocs}
  \link{groupDocInfo} \link{groupPeople} \link{leaveGroup}
  \link{unfollowGroup} \link{getGroupID}

  \strong{Useful links}
  \href{http://mendeley.com/}{Mendeley}
  \href{http://apidocs.mendeley.com/}{Full API
  documentation}
}

